requirements.txt
fire==0.7.1
tqdm==4.67.1
click==8.2.1
srsly==2.5.1
sympy==1.13.1
PyYAML==6.0.2
psutil==7.0.0
pynvml==12.0.0
tenacity==9.1.2
shortuuid==1.0.13
omegaconf==2.3.0
hydra-core==1.3.2
setuptools==78.1.1
class_registry==2.1.2
timeout_decorator==0.5.0
antlr4-python3-runtime==4.9.3

wandb==0.19.8
pandas==2.3.2
sacrebleu==2.5.1
word2number==1.1
pylatexenc==2.10
math_verify==0.8.0
tensorboard==2.20.0
latex2sympy2==1.5.4
wikipedia_api==0.8.1
beautifulsoup4==4.13.4

peft==0.17.1
torch==2.6.0
datasets==3.6.0
deepspeed==0.16.8
vllm==0.8.5.post1
transformers==4.56.0
ray[default]==2.43.0
ring_flash_attn==0.1.7
flash_attn==2.7.0.post2
